move default_data and related to ParametersBase
- make ParametersBase abstract base class
- test that Growth and Behavior now also support default_data
